Merged PR 3957: Deploying Leo programs in pipeline
Created a deployment pipeline that also deploys Leo programs to the testnet. It does the following:

1. Checks if any programs are updated.
2. Deploys the updated programs to the Aleo testnet.
3. Stores the names of the new programs.
4. Builds the API.
5. Sets the program names as env variables for the API and runs the API.

Note: Files in `deployment-old` are not being used, but can be useful for future reference.

Also created a README that explains how to run the API and what to do before completing a PR.

Related work items: #17950, #17951, #17952, #17953, #17954